requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pytest==8.3.3
asyncpg==0.29.0
pgvector==0.3.0
sentence-transformers==3.1.1
streamlit==1.39.0
llama-index==0.11.17
python-dotenv==1.0.1
psycopg2-binary==2.9.9
